Added "chartemp"functionality
chartemp is a console command that allows us to put a player on a temporary/template character.

A temporary character can either copy the properties of an existing character of another player, or it can copy a character from a template file.

Being on a temporary character blocks your access to /skin and /nick
(at least for the time being)